+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.0.5] - 2024-08-22
+
+### Added
+
+- Added DatePickerWidget, DateRangePickerWidget, TimeSpanPickerWidget
+- Added DateRangeFormField, TimeSpanFormField
+
+### Changed
+
+- DateFormField now uses the DatePickerWidget
+
 ## [1.0.4] - 2024-08-20
 
 ### Added